# openapi/openapi.yaml
...
openapi: 3.0.2
info:
  title: Giveback Ninja
  version: 2.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
paths:
...
  /history/{period}:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Period'
              examples:
                fy20:
                  value:
                    cumulatedOn: 2021-01-11T08:08:56.5899687
                    records:
                      retired_ninja:
                        score: 120
                        level: BLACK
                      super_ninja:
                        score: 80
                        level: BLACK
                      modest_ninja:
                        score: 30
                        level: GREY
                    name: FY20
          description: Period requested with path parameter
        "404":
          $ref: '#/components/responses/NotFoundResponse'
      summary: Obtains history for requested period
    post:
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Period'
              examples:
                fy20:
                  value:
                    cumulatedOn: 2021-06-11T08:08:56.5899687
                    records:
                      retired_ninja:
                        score: 120
                        level: BLACK
                      super_ninja:
                        score: 80
                        level: BLACK
                      modest_ninja:
                        score: 30
                        level: GREY
                    name: FY20
          description: Created Period
      summary: Generates a new Historic period and reset existing scores
    parameters:
    - examples:
        fy20:
          value: FY20
      name: period
      description: Period name
      schema:
        type: string
      in: path
      required: true
components:
  schemas:
...
    Record:
      description: ""
      required:
      - level
      - score
      type: object
      properties:
        level:
          description: The level attained by the user for a period
          type: string
        score:
          description: Points cumulated by a user during a period
          type: integer
    Period:
      description: The period for which the history was computed. Usually Fiscal Year
      required:
      - name
      - cumulatedOn
      - records
      type: object
      properties:
        name:
          description: Unique name of the period
          type: string
        cumulatedOn:
          format: date-time
          description: When the period was computed
          type: string
        records:
          description: The records for this period
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Record'
...
    ErrorResponse:
      title: Root Type for ErrorResponse
      description: An error response
      type: object
      properties:
        message:
          description: The error message
          type: string
      example:
        message: Entity was not found
...
  responses:
    NotFoundResponse:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            entity-not-found:
              value:
                message: Entity was not found
      description: ""